Bug 738040: replace custom PATH with links in /usr/local/bin; r=kmoir
This upgrades some RPMs:
    git.spec - links added
    mock_mozilla.spec - not in /tools
    py27_mercurial.spec - links added
    py27_virtualenv.spec - links added
    python26.spec - links added
    python27.spec - links added
    tooltool.pp - no package
and adds /usr/local/bin to PATH only for the builder user.  It
specifically removes the global PATH-setting file.

(I went ahead and added the links in python26, as it was the only unused
package - I'm sorry to change my mind so much!)

Note that this does *not* add a 'python' link, so with or without this
patch, a bare 'python' will run the system python (2.6.6).