Bug 779097: add anchors to users:*
This adds anchors so that the action of users::xx is completely
encapsulated by Class['users::xx'] in a require or before statement.

Actually, that's not quite true.  The user accounts themselves are
created in a new 'users' stage that runs before 'main'.  This avoids
problems where the automatic user dependency doesn't work (e.g., the
temporary darwinuser type).