Bug 1014047: use directory environments, with a long environment_timeout; r=Callek
This is a re-landing, with environment_timeout changed from 10s to 600s
on the hypothesis that the short timeout was behind the performance
issues.

Note that this makes existing user environments invalid, as they are
located at /etc/puppet/environments/$user/env/, and the "env/" suffix
is not compatible with directory environments.